+++ packages/york-web/src/components/complex/Header/index.js
@@ -54,7 +54,7 @@ Header.propTypes = {
   components: PropTypes.objectOf(PropTypes.elementType.isRequired),
   content: PropTypes.shape({
     phone: PropTypes.string,
-    // /// proper shape
+    /// proper shape
     regions: PropTypes.array.isRequired,
     tabs: PropTypes.arrayOf(menuItemShape.isRequired).isRequired,
     profile: PropTypes.arrayOf(menuItemShape.isRequired).isRequired,
